Issue 2489 Refactoring: group processing-related classes in separate files
