+++ deploy/docker/cp-tools/research/hcs-parser/parser/process_hcs_files_cluster.py
@@ -1,4 +1,4 @@
-# Copyright 2022 EPAM Systems, Inc. (https://www.epam.com/)
+# Copyright 2017-2022 EPAM Systems, Inc. (https://www.epam.com/)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,203 +13,33 @@
 # limitations under the License.
 
 import os
-import datetime
 import math
 import tempfile
 import time
-import xml.etree.ElementTree as ET
 import multiprocessing
 import re
 import traceback
 import subprocess
-from pipeline.api import PipelineAPI, TaskStatus
-from pipeline.log import Logger
+
+from .src.fs import get_processing_roots
+from .src.utils import HcsFileLogger, log_info, log_success
+from .src.utils import get_int_run_param, get_bool_run_param
 
 
-def get_int_env_var(env_var_name, default_value):
-    return int(os.getenv(env_var_name, default_value))
+CLUSTER_MAX_SIZE = get_int_run_param('CP_CAP_AUTOSCALE_WORKERS', 1)
+TAGS_PROCESSING_ONLY = get_bool_run_param('HCS_PARSING_TAGS_ONLY')
+FORCE_PROCESSING = get_bool_run_param('HCS_FORCE_PROCESSING')
 
-
-def log_success(message):
-    log_info(message, status=TaskStatus.SUCCESS)
-
-
-def log_info(message, status=TaskStatus.RUNNING):
-    Logger.log_task_event(HCS_PROCESSING_TASK_NAME, message, status)
-
-
-LOOKUP_PATHS = os.getenv('HCS_LOOKUP_DIRECTORIES', '').split(',')
-CLUSTER_MAX_SIZE = get_int_env_var('CP_CAP_AUTOSCALE_WORKERS', 1)
 RUN_ID = os.getenv('RUN_ID')
 
-HCS_ACTIVE_PROCESSING_TIMEOUT_MIN = get_int_env_var('HCS_ACTIVE_PROCESSING_TIMEOUT_MIN', 360)
-TAGS_PROCESSING_ONLY = os.getenv('HCS_PARSING_TAGS_ONLY', 'false') == 'true'
-FORCE_PROCESSING = os.getenv('HCS_FORCE_PROCESSING', 'false') == 'true'
-HCS_CLOUD_FILES_SCHEMA = os.getenv('HCS_CLOUD_FILES_SCHEMA', 's3')
-
-HCS_PROCESSING_OUTPUT_FOLDER = os.getenv('HCS_PARSING_OUTPUT_FOLDER')
-HCS_PROCESSING_TASK_NAME = 'HCS processing'
-HCS_OME_COMPATIBLE_INDEX_FILE_NAME = 'Index.xml'
 HCS_INDEX_FILE_NAME = os.getenv('HCS_PARSING_INDEX_FILE_NAME', 'Index.xml')
 HCS_IMAGE_DIR_NAME = os.getenv('HCS_PARSING_IMAGE_DIR_NAME', 'Images')
-HCS_CLUSTER_PROCESSING_MEMORY_SIZE_SLOT_FACTOR = get_int_env_var('HCS_PARSING_CLUSTER_PROCESSING_MEMORY_FACTOR', 20)
-HCS_CLUSTER_PROCESSING_MEMORY_CLUSTER_SLOT = \
-    get_int_env_var('HCS_PARSING_CLUSTER_PROCESSING_MEMORY_PER_CLUSTER_SLOT', 16)
-HCS_CLUSTER_PROCESSING_MEMORY_INSTANCE_SLOT = \
-    get_int_env_var('HCS_PARSING_CLUSTER_PROCESSING_MEMORY_PER_INSTANCE_SLOT', 8)
 MEASUREMENT_INDEX_FILE_PATH = '/{}/{}'.format(HCS_IMAGE_DIR_NAME, HCS_INDEX_FILE_NAME)
-
-
-class HcsParsingUtils:
-
-    @staticmethod
-    def extract_xml_schema(xml_info_root):
-        full_schema = xml_info_root.tag
-        return full_schema[:full_schema.rindex('}') + 1]
-
-    @staticmethod
-    def get_file_without_extension(file_path):
-        return os.path.splitext(file_path)[0]
-
-    @staticmethod
-    def get_basename_without_extension(file_path):
-        return HcsParsingUtils.get_file_without_extension(os.path.basename(file_path))
-
-    @staticmethod
-    def get_file_last_modification_time(file_path):
-        return int(os.stat(file_path).st_mtime)
-
-    @staticmethod
-    def build_preview_file_path(hcs_root_folder_path):
-        index_file_abs_path = os.path.join(HcsParsingUtils.get_file_without_extension(hcs_root_folder_path),
-                                           HCS_IMAGE_DIR_NAME, HCS_INDEX_FILE_NAME)
-        hcs_xml_info_root = ET.parse(index_file_abs_path).getroot()
-        hcs_schema_prefix = HcsParsingUtils.extract_xml_schema(hcs_xml_info_root)
-        file_name = HcsParsingUtils.get_file_without_extension(hcs_root_folder_path)
-        name_xml_element = HcsParsingUtils.extract_plate_from_hcs_xml(hcs_xml_info_root, hcs_schema_prefix) \
-            .find(hcs_schema_prefix + 'Name')
-        if name_xml_element is not None:
-            file_pretty_name = name_xml_element.text
-            if file_pretty_name is not None:
-                file_name = file_pretty_name
-        preview_file_basename = file_name + '.hcs'
-        parent_folder = HCS_PROCESSING_OUTPUT_FOLDER \
-            if HCS_PROCESSING_OUTPUT_FOLDER is not None \
-            else os.path.dirname(hcs_root_folder_path)
-        return os.path.join(parent_folder, preview_file_basename)
-
-    @staticmethod
-    def extract_plate_from_hcs_xml(hcs_xml_info_root, hcs_schema_prefix=None):
-        if not hcs_schema_prefix:
-            hcs_schema_prefix = HcsParsingUtils.extract_xml_schema(hcs_xml_info_root)
-        plates_list = hcs_xml_info_root.find(hcs_schema_prefix + 'Plates')
-        plate = plates_list.find(hcs_schema_prefix + 'Plate')
-        return plate
-
-    @staticmethod
-    def get_stat_active_file_name(hcs_img_path):
-        return HcsParsingUtils._get_service_file_name(hcs_img_path, 'hcsparser.inprog')
-
-    @staticmethod
-    def get_stat_file_name(hcs_img_path):
-        return HcsParsingUtils._get_service_file_name(hcs_img_path, 'hcsparser')
-
-    @staticmethod
-    def get_service_directory(hcs_img_path):
-        name_without_extension = HcsParsingUtils.get_basename_without_extension(hcs_img_path)
-        parent_dir = HCS_PROCESSING_OUTPUT_FOLDER \
-            if HCS_PROCESSING_OUTPUT_FOLDER is not None \
-            else os.path.dirname(hcs_img_path)
-        return os.path.join(parent_dir, '.hcsparser', name_without_extension)
-
-    @staticmethod
-    def generate_local_service_directory(hcs_img_path):
-        name_without_extension = HcsParsingUtils.get_basename_without_extension(hcs_img_path)
-        return tempfile.mkdtemp(prefix=name_without_extension + '.hcsparser.')
-
-    @staticmethod
-    def create_service_dir_if_not_exist(hcs_img_path):
-        directory = HcsParsingUtils.get_service_directory(hcs_img_path)
-        if not os.path.exists(directory):
-            os.makedirs(directory)
-
-    @staticmethod
-    def _get_service_file_name(hcs_img_path, suffix):
-        parent_dir = HcsParsingUtils.get_service_directory(hcs_img_path)
-        parser_flag_file = '.stat.{}'.format(suffix)
-        return os.path.join(parent_dir, parser_flag_file)
-
-    @staticmethod
-    def active_processing_exceed_timeout(active_stat_file):
-        processing_stat_file_modification_date = HcsParsingUtils.get_file_last_modification_time(active_stat_file)
-        processing_deadline = datetime.datetime.now() - datetime.timedelta(minutes=HCS_ACTIVE_PROCESSING_TIMEOUT_MIN)
-        return (processing_stat_file_modification_date - time.mktime(processing_deadline.timetuple())) < 0
-
-    @staticmethod
-    def extract_cloud_path(file_path, cloud_scheme=HCS_CLOUD_FILES_SCHEMA):
-        path_chunks = file_path.split('/cloud-data/', 1)
-        if len(path_chunks) != 2:
-            raise RuntimeError('Unable to determine cloud path of [{}]'.format(file_path))
-        return '{}://{}'.format(cloud_scheme, path_chunks[1])
-
-    @staticmethod
-    def quote_string(string):
-        return '"{}"'.format(string)
-
-
-class HcsFileLogger:
-
-    def __init__(self, file_path):
-        self.file_path = file_path
-
-    def log_info(self, message, status=TaskStatus.RUNNING):
-        Logger.log_task_event(HCS_PROCESSING_TASK_NAME, '[{}] {}'.format(self.file_path, message), status)
-
-
-class HcsProcessingDirsGenerator:
-
-    def __init__(self, lookup_paths):
-        self.lookup_paths = lookup_paths
-
-    @staticmethod
-    def is_folder_content_modified_after(dir_path, modification_date):
-        dir_root = os.walk(dir_path)
-        for dir_root, directories, files in dir_root:
-            for file in files:
-                if HcsParsingUtils.get_file_last_modification_time(os.path.join(dir_root, file)) > modification_date:
-                    return True
-        return False
-
-    def generate_paths(self):
-        hcs_roots = self.find_all_hcs_roots()
-        log_info('Found {} HCS files'.format(len(hcs_roots)))
-        return filter(lambda p: self.is_processing_required(p), hcs_roots)
-
-    def find_all_hcs_roots(self):
-        hcs_roots = set()
-        for lookup_path in self.lookup_paths:
-            dir_walk_root = os.walk(lookup_path)
-            for dir_root, directories, files in dir_walk_root:
-                for file in files:
-                    full_file_path = os.path.join(dir_root, file)
-                    if full_file_path.endswith(MEASUREMENT_INDEX_FILE_PATH):
-                        hcs_roots.add(full_file_path[:-len(MEASUREMENT_INDEX_FILE_PATH)])
-        return hcs_roots
-
-    def is_processing_required(self, hcs_folder_root_path):
-        if TAGS_PROCESSING_ONLY or FORCE_PROCESSING:
-            return True
-        hcs_img_path = HcsParsingUtils.build_preview_file_path(hcs_folder_root_path)
-        if not os.path.exists(hcs_img_path):
-            return True
-        active_stat_file = HcsParsingUtils.get_stat_active_file_name(hcs_img_path)
-        if os.path.exists(active_stat_file):
-            return HcsParsingUtils.active_processing_exceed_timeout(active_stat_file)
-        stat_file = HcsParsingUtils.get_stat_file_name(hcs_img_path)
-        if not os.path.isfile(stat_file):
-            return True
-        stat_file_modification_date = HcsParsingUtils.get_file_last_modification_time(stat_file)
-        return self.is_folder_content_modified_after(hcs_folder_root_path, stat_file_modification_date)
+HCS_CLUSTER_PROCESSING_MEMORY_SIZE_SLOT_FACTOR = get_int_run_param('HCS_PARSING_CLUSTER_PROCESSING_MEMORY_FACTOR', 20)
+HCS_CLUSTER_PROCESSING_MEMORY_CLUSTER_SLOT = \
+    get_int_run_param('HCS_PARSING_CLUSTER_PROCESSING_MEMORY_PER_CLUSTER_SLOT', 16)
+HCS_CLUSTER_PROCESSING_MEMORY_INSTANCE_SLOT = \
+    get_int_run_param('HCS_PARSING_CLUSTER_PROCESSING_MEMORY_PER_INSTANCE_SLOT', 8)
 
 
 class HcsFileSgeParser:
@@ -320,9 +150,9 @@ class HcsFileSgeParser:
     def _extract_job_stats(self, job_id):
         job_stats_command = 'qacct -j ' + job_id
         self.log_info('Extracting execution stats for jobId=[{}]'.format(job_id))
-        stats_extraction_retry_delay = get_int_env_var('HCS_CLUSTER_PROCESSING_QACCT_DELAY', 15)
+        stats_extraction_retry_delay = get_int_run_param('HCS_CLUSTER_PROCESSING_QACCT_DELAY', 15)
         time.sleep(stats_extraction_retry_delay)
-        retries = get_int_env_var('HCS_CLUSTER_PROCESSING_QACCT_RETRIES', 5)
+        retries = get_int_run_param('HCS_CLUSTER_PROCESSING_QACCT_RETRIES', 5)
         while retries > 0:
             try:
                 return subprocess.check_output(job_stats_command, shell=True)
@@ -385,18 +215,14 @@ def try_process_hcs_in_cluster(hcs_root_dir):
 
 
 def process_hcs_files_cluster():
-    if not LOOKUP_PATHS:
-        log_success('No paths for HCS processing specified')
-        exit(0)
-    log_info('Following paths are specified for processing: {}'.format(LOOKUP_PATHS))
-    log_info('Lookup for unprocessed files')
-    paths_to_hcs_files = HcsProcessingDirsGenerator(LOOKUP_PATHS).generate_paths()
-    if not paths_to_hcs_files:
+    should_force_processing = TAGS_PROCESSING_ONLY or FORCE_PROCESSING
+    paths_to_hcs_roots = get_processing_roots(should_force_processing, MEASUREMENT_INDEX_FILE_PATH)
+    if not paths_to_hcs_roots or len(paths_to_hcs_roots) == 0:
         log_success('Found no files requires processing in the lookup directories.')
         exit(0)
-    log_info('Found {} files for processing.'.format(len(paths_to_hcs_files)))
+    log_info('Found {} files for processing.'.format(len(paths_to_hcs_roots)))
     pool = multiprocessing.Pool(CLUSTER_MAX_SIZE)
-    pool.map(try_process_hcs_in_cluster, paths_to_hcs_files)
+    pool.map(try_process_hcs_in_cluster, paths_to_hcs_roots)
     log_success('Finished HCS files processing')
     exit(0)
