+++ deploy/docker/cp-tools/research/hcs-parser/parser/src/processors.py
@@ -1,4 +1,4 @@
-# Copyright 2022 EPAM Systems, Inc. (https://www.epam.com/)
+# Copyright 2017-2022 EPAM Systems, Inc. (https://www.epam.com/)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,15 +15,12 @@
 import errno
 import json
 import os
-import multiprocessing
 import datetime
 import math
 import shutil
 import sys
 import tempfile
-import time
 import traceback
-import urllib
 import xml.etree.ElementTree as ET
 
 from collections import OrderedDict
@@ -33,181 +30,20 @@ from pipeline.log import Logger
 from pipeline.common import get_path_with_trailing_delimiter, get_path_without_trailing_delimiter
 from PIL import Image
 
+from .utils import HcsParsingUtils, log_info, HCS_PROCESSING_TASK_NAME
+from .tags import HcsFileTagProcessor
 
-MAGNIFICATION_CAT_ATTR_NAME = os.getenv('HCS_PARSING_MAGNIFICATION_CAT_ATTR_NAME', 'Magnification')
-HCS_ACTIVE_PROCESSING_TIMEOUT_MIN = int(os.getenv('HCS_PARSING_ACTIVE_PROCESSING_TIMEOUT_MIN', 360))
 TAGS_PROCESSING_ONLY = os.getenv('HCS_PARSING_TAGS_ONLY', 'false') == 'true'
-FORCE_PROCESSING = os.getenv('HCS_PARSING_FORCE_PROCESSING', 'false') == 'true'
-HCS_CLOUD_FILES_SCHEMA = os.getenv('HCS_PARSING_CLOUD_FILES_SCHEMA', 's3')
 PLANE_COORDINATES_DELIMITER = os.getenv('HCS_PARSING_PLANE_COORDINATES_DELIMITER', '_')
 DEFAULT_CHANNEL_WIDTH = os.getenv('HCS_PARSING_DEFAULT_CHANNEL_WIDTH', 1080)
 DEFAULT_CHANNEL_HEIGHT = os.getenv('HCS_PARSING_DEFAULT_CHANNEL_HEIGHT', 1080)
 
-HCS_PROCESSING_OUTPUT_FOLDER = os.getenv('HCS_PARSING_OUTPUT_FOLDER')
-HCS_PROCESSING_TASK_NAME = 'HCS processing'
 HCS_OME_COMPATIBLE_INDEX_FILE_NAME = 'Index.xml'
 OVERVIEW_DIR_NAME = 'overview'
 HCS_INDEX_FILE_NAME = os.getenv('HCS_PARSING_INDEX_FILE_NAME', 'Index.xml')
 HCS_IMAGE_DIR_NAME = os.getenv('HCS_PARSING_IMAGE_DIR_NAME', 'Images')
 MEASUREMENT_INDEX_FILE_PATH = '/{}/{}'.format(HCS_IMAGE_DIR_NAME, HCS_INDEX_FILE_NAME)
-UNKNOWN_ATTRIBUTE_VALUE = 'NA'
-HYPHEN = '-'
-COMMA = ','
-TAGS_MAPPING_KEYS_DELIMITER = '='
 PATH_DELIMITER = '/'
-
-HCS_KEYWORD_FILE_SUFFIX = '.kw.txt'
-EXTENDED_TAGS_OWNER_TAG_KEY = 'OWNER'
-EXTENDED_TAGS_CHANNEL_TYPE_TAG_KEY = 'CHANNELTYPE'
-EXTENDED_TAGS_MULTIPLANE_FLAG_TAG_KEY = 'PLANES'
-EXTENDED_TAGS_TIMECOURSE_FLAG_TAG_KEY = 'TIMEPOINTS'
-
-
-def log_success(message):
-    log_info(message, status=TaskStatus.SUCCESS)
-
-
-def log_info(message, status=TaskStatus.RUNNING):
-    Logger.log_task_event(HCS_PROCESSING_TASK_NAME, message, status)
-
-
-class HcsParsingUtils:
-
-    @staticmethod
-    def extract_xml_schema(xml_info_root):
-        full_schema = xml_info_root.tag
-        return full_schema[:full_schema.rindex('}') + 1]
-
-    @staticmethod
-    def get_file_without_extension(file_path):
-        return os.path.splitext(file_path)[0]
-
-    @staticmethod
-    def get_basename_without_extension(file_path):
-        return HcsParsingUtils.get_file_without_extension(os.path.basename(file_path))
-
-    @staticmethod
-    def get_file_last_modification_time(file_path):
-        return int(os.stat(file_path).st_mtime)
-
-    @staticmethod
-    def build_preview_file_path(hcs_root_folder_path):
-        index_file_abs_path = os.path.join(HcsParsingUtils.get_file_without_extension(hcs_root_folder_path),
-                                           HCS_IMAGE_DIR_NAME, HCS_INDEX_FILE_NAME)
-        hcs_xml_info_root = ET.parse(index_file_abs_path).getroot()
-        hcs_schema_prefix = HcsParsingUtils.extract_xml_schema(hcs_xml_info_root)
-        file_name = HcsParsingUtils.get_file_without_extension(hcs_root_folder_path)
-        name_xml_element = HcsFileParser.extract_plate_from_hcs_xml(hcs_xml_info_root, hcs_schema_prefix)\
-            .find(hcs_schema_prefix + 'Name')
-        if name_xml_element is not None:
-            file_pretty_name = name_xml_element.text
-            if file_pretty_name is not None:
-                file_name = file_pretty_name
-        preview_file_basename = HcsParsingUtils.replace_special_chars(file_name) + '.hcs'
-        parent_folder = HCS_PROCESSING_OUTPUT_FOLDER \
-            if HCS_PROCESSING_OUTPUT_FOLDER is not None \
-            else os.path.dirname(hcs_root_folder_path)
-        return os.path.join(parent_folder, preview_file_basename)
-
-    @staticmethod
-    def get_stat_active_file_name(hcs_img_path):
-        return HcsParsingUtils._get_service_file_name(hcs_img_path, 'hcsparser.inprog')
-
-    @staticmethod
-    def get_stat_file_name(hcs_img_path):
-        return HcsParsingUtils._get_service_file_name(hcs_img_path, 'hcsparser')
-
-    @staticmethod
-    def get_service_directory(hcs_img_path):
-        name_without_extension = HcsParsingUtils.get_basename_without_extension(hcs_img_path)
-        parent_dir = HCS_PROCESSING_OUTPUT_FOLDER \
-            if HCS_PROCESSING_OUTPUT_FOLDER is not None \
-            else os.path.dirname(hcs_img_path)
-        return os.path.join(parent_dir, '.hcsparser', name_without_extension)
-
-    @staticmethod
-    def generate_local_service_directory(hcs_img_path):
-        name_without_extension = HcsParsingUtils.get_basename_without_extension(hcs_img_path)
-        return tempfile.mkdtemp(prefix=name_without_extension + '.hcsparser.')
-
-    @staticmethod
-    def create_service_dir_if_not_exist(hcs_img_path):
-        directory = HcsParsingUtils.get_service_directory(hcs_img_path)
-        if not os.path.exists(directory):
-            os.makedirs(directory)
-
-    @staticmethod
-    def _get_service_file_name(hcs_img_path, suffix):
-        parent_dir = HcsParsingUtils.get_service_directory(hcs_img_path)
-        parser_flag_file = '.stat.{}'.format(suffix)
-        return os.path.join(parent_dir, parser_flag_file)
-
-    @staticmethod
-    def active_processing_exceed_timeout(active_stat_file):
-        processing_stat_file_modification_date = HcsParsingUtils.get_file_last_modification_time(active_stat_file)
-        processing_deadline = datetime.datetime.now() - datetime.timedelta(minutes=HCS_ACTIVE_PROCESSING_TIMEOUT_MIN)
-        return (processing_stat_file_modification_date - time.mktime(processing_deadline.timetuple())) < 0
-
-    @staticmethod
-    def extract_cloud_path(file_path, cloud_scheme=HCS_CLOUD_FILES_SCHEMA):
-        path_chunks = file_path.split('/cloud-data/', 1)
-        if len(path_chunks) != 2:
-            raise RuntimeError('Unable to determine cloud path of [{}]'.format(file_path))
-        return '{}://{}'.format(cloud_scheme, path_chunks[1])
-
-    @staticmethod
-    def replace_special_chars(file_path):
-        return file_path.replace(PATH_DELIMITER, '|')
-
-    @staticmethod
-    def quote_string(string):
-        return '"{}"'.format(string)
-
-
-class HcsProcessingDirsGenerator:
-
-    def __init__(self, lookup_paths):
-        self.lookup_paths = lookup_paths
-
-    @staticmethod
-    def is_folder_content_modified_after(dir_path, modification_date):
-        dir_root = os.walk(dir_path)
-        for dir_root, directories, files in dir_root:
-            for file in files:
-                if HcsParsingUtils.get_file_last_modification_time(os.path.join(dir_root, file)) > modification_date:
-                    return True
-        return False
-
-    def generate_paths(self):
-        hcs_roots = self.find_all_hcs_roots()
-        log_info('Found {} HCS files'.format(len(hcs_roots)))
-        return filter(lambda p: self.is_processing_required(p), hcs_roots)
-
-    def find_all_hcs_roots(self):
-        hcs_roots = set()
-        for lookup_path in self.lookup_paths:
-            dir_walk_root = os.walk(lookup_path)
-            for dir_root, directories, files in dir_walk_root:
-                for file in files:
-                    full_file_path = os.path.join(dir_root, file)
-                    if full_file_path.endswith(MEASUREMENT_INDEX_FILE_PATH):
-                        hcs_roots.add(full_file_path[:-len(MEASUREMENT_INDEX_FILE_PATH)])
-        return hcs_roots
-
-    def is_processing_required(self, hcs_folder_root_path):
-        if TAGS_PROCESSING_ONLY or FORCE_PROCESSING:
-            return True
-        hcs_img_path = HcsParsingUtils.build_preview_file_path(hcs_folder_root_path)
-        if not os.path.exists(hcs_img_path):
-            return True
-        active_stat_file = HcsParsingUtils.get_stat_active_file_name(hcs_img_path)
-        if os.path.exists(active_stat_file):
-            return HcsParsingUtils.active_processing_exceed_timeout(active_stat_file)
-        stat_file = HcsParsingUtils.get_stat_file_name(hcs_img_path)
-        if not os.path.isfile(stat_file):
-            return True
-        stat_file_modification_date = HcsParsingUtils.get_file_last_modification_time(stat_file)
-        return self.is_folder_content_modified_after(hcs_folder_root_path, stat_file_modification_date)
 
 
 class FieldDetails:
@@ -238,207 +74,6 @@ class WellGrid:
 
     def get_values_dict(self):
         return dict({y_coord: set(self.__x_coords) for y_coord in self.__y_coords})
-
-
-class HcsFileTagProcessor:
-
-    CATEGORICAL_ATTRIBUTE = '/categoricalAttribute'
-
-    def __init__(self, hcs_root_dir, hcs_img_path, xml_info_tree):
-        self.hcs_root_dir = hcs_root_dir
-        self.hcs_img_path = hcs_img_path
-        self.xml_info_tree = xml_info_tree
-        self.api = PipelineAPI(os.environ['API'], 'logs')
-        self.system_dictionaries_url = self.api.api_url + self.CATEGORICAL_ATTRIBUTE
-        self.cloud_path = HcsParsingUtils.extract_cloud_path(self.hcs_img_path)
-
-    @staticmethod
-    def build_magnification_from_numeric_string(magnification):
-        return '{}x'.format(int(float(magnification)))
-
-    @staticmethod
-    def map_to_metadata_dict(schema, metadata):
-        metadata_entries = metadata.findall(schema + 'XMLAnnotation')
-        metadata_dict = dict()
-        for entry in metadata_entries:
-            entry_value = entry.find(schema + 'Value')
-            if entry_value is not None:
-                metadata_record = entry_value.find(schema + 'OriginalMetadata')
-                if metadata_record is not None:
-                    key = metadata_record.find(schema + 'Key').text
-                    value = metadata_record.find(schema + 'Value').text
-                    if key and value:
-                        metadata_dict[key] = value
-        return metadata_dict
-
-    def log_processing_info(self, message, status=TaskStatus.RUNNING):
-        log_info('[{}] {}'.format(self.hcs_root_dir, message), status=status)
-
-    def process_tags(self):
-        existing_attributes_dictionary = self.load_existing_attributes()
-        ome_schema = HcsParsingUtils.extract_xml_schema(self.xml_info_tree)
-        metadata = self.xml_info_tree.find(ome_schema + 'StructuredAnnotations')
-        if not metadata:
-            self.log_processing_info('No metadata found for file, skipping tags processing...')
-            return 0
-        metadata_dict = self.map_to_metadata_dict(ome_schema, metadata)
-        self.add_metadata_from_keyword_file(metadata_dict)
-        tags_to_push = self.build_tags_dictionary(metadata_dict, existing_attributes_dictionary)
-        if not tags_to_push:
-            self.log_processing_info('No matching tags found')
-            return 0
-        pipe_tags = self.prepare_tags(existing_attributes_dictionary, tags_to_push)
-        tags_to_push_str = ' '.join(pipe_tags)
-        self.log_processing_info('Following tags will be assigned to the file: {}'.format(tags_to_push_str))
-        url_encoded_cloud_file_path = HcsParsingUtils.extract_cloud_path(urllib.quote(self.hcs_img_path))
-        return os.system('pipe storage set-object-tags "{}" {}'.format(url_encoded_cloud_file_path, tags_to_push_str))
-
-    def add_metadata_from_keyword_file(self, metadata_dict):
-        experiment_id = os.path.basename(self.hcs_root_dir)
-        keyword_file_name = experiment_id[:experiment_id.rfind(HYPHEN)] + HCS_KEYWORD_FILE_SUFFIX
-        keyword_file_path = os.path.join(self.hcs_root_dir, keyword_file_name)
-        keyword_file_content = self.try_extract_keyword_file_content(keyword_file_path)
-        if keyword_file_content is None:
-            return
-        self.add_owner_info_if_present(keyword_file_content, metadata_dict)
-        self.add_channel_type_info_if_present(keyword_file_content, metadata_dict)
-        self.add_timepoints_info_if_present(keyword_file_content, metadata_dict)
-        self.add_plane_info_if_present(keyword_file_content, metadata_dict)
-
-    @staticmethod
-    def try_extract_keyword_file_content(keyword_file_path):
-        if os.path.isfile(keyword_file_path):
-            with open(keyword_file_path, "r") as keyword_file:
-                data = keyword_file.read()
-                json_data = data[data.find('{'):data.rfind('}') + 1]
-                return json.loads(json_data)
-        return None
-
-    @staticmethod
-    def add_plane_info_if_present(keyword_file_content, metadata_dict):
-        if EXTENDED_TAGS_MULTIPLANE_FLAG_TAG_KEY in keyword_file_content:
-            planes_count = int(keyword_file_content[EXTENDED_TAGS_MULTIPLANE_FLAG_TAG_KEY])
-            metadata_dict[EXTENDED_TAGS_MULTIPLANE_FLAG_TAG_KEY] = 'Yes' if planes_count > 1 else 'No'
-
-    @staticmethod
-    def add_timepoints_info_if_present(keyword_file_content, metadata_dict):
-        if EXTENDED_TAGS_TIMECOURSE_FLAG_TAG_KEY in keyword_file_content:
-            timepoints_count = int(keyword_file_content[EXTENDED_TAGS_TIMECOURSE_FLAG_TAG_KEY])
-            metadata_dict[EXTENDED_TAGS_TIMECOURSE_FLAG_TAG_KEY] = 'Yes' if timepoints_count > 1 else 'No'
-
-    @staticmethod
-    def add_channel_type_info_if_present(keyword_file_content, metadata_dict):
-        if EXTENDED_TAGS_CHANNEL_TYPE_TAG_KEY in keyword_file_content:
-            channel_type = keyword_file_content[EXTENDED_TAGS_CHANNEL_TYPE_TAG_KEY]
-            if isinstance(channel_type, list):
-                channel_type = PATH_DELIMITER.join(channel_type.sort())
-            metadata_dict[EXTENDED_TAGS_CHANNEL_TYPE_TAG_KEY] = channel_type
-
-    @staticmethod
-    def add_owner_info_if_present(keyword_file_content, metadata_dict):
-        if EXTENDED_TAGS_OWNER_TAG_KEY in keyword_file_content:
-            metadata_dict[EXTENDED_TAGS_OWNER_TAG_KEY] = keyword_file_content[EXTENDED_TAGS_OWNER_TAG_KEY].upper()
-
-    def build_tags_dictionary(self, metadata_dict, existing_attributes_dictionary):
-        tags_dictionary = dict()
-        common_tags_mapping = self.map_tags(existing_attributes_dictionary)
-        if common_tags_mapping:
-            tags_dictionary.update(self.extract_matching_tags_from_metadata(metadata_dict, common_tags_mapping))
-        magnification_value = self._get_magnification_attribute_value()
-        if magnification_value:
-            tags_dictionary[MAGNIFICATION_CAT_ATTR_NAME] = {magnification_value}
-        return tags_dictionary
-
-    def _get_magnification_attribute_value(self):
-        ome_schema = HcsParsingUtils.extract_xml_schema(self.xml_info_tree)
-        instrument_details = self.xml_info_tree.find(ome_schema + 'Instrument')
-        if len(instrument_details) > 0:
-            objectives_details = instrument_details.find(ome_schema + 'Objective')
-            if objectives_details is not None:
-                magnification_value = objectives_details.get('NominalMagnification')
-                if magnification_value:
-                    return self.build_magnification_from_numeric_string(magnification_value)
-        return None
-
-    def prepare_tags(self, existing_attributes_dictionary, tags_to_push):
-        attribute_updates = list()
-        pipe_tags = list()
-        for attribute_name, values_to_push in tags_to_push.items():
-            if len(values_to_push) > 1:
-                self.log_processing_info('Multiple tags matches occurred for "{}": [{}]'
-                                         .format(attribute_name, values_to_push))
-                continue
-            value = list(values_to_push)[0]
-            if attribute_name not in existing_attributes_dictionary:
-                self.log_processing_info('No categorical attribute [{}] exists'.format(attribute_name))
-                continue
-            existing_values = existing_attributes_dictionary[attribute_name]
-            existing_attribute_id = existing_values[0]['attributeId']
-            existing_values_names = [existing_value['value'] for existing_value in existing_values]
-            if value not in existing_values_names:
-                existing_values.append({'key': attribute_name, 'value': value})
-                attribute_updates.append({'id': int(existing_attribute_id),
-                                          'key': attribute_name, 'values': existing_values})
-            pipe_tags.append('\'{}\'=\'{}\''.format(attribute_name, value))
-        if attribute_updates:
-            self.log_processing_info('Updating following categorical attributes before tagging: {}'
-                                     .format(attribute_updates))
-            self.update_categorical_attributes(attribute_updates)
-        return pipe_tags
-
-    def update_categorical_attributes(self, attribute_updates):
-        for attribute in attribute_updates:
-            self.api.execute_request(self.system_dictionaries_url, method='post', data=json.dumps(attribute))
-
-    def extract_matching_tags_from_metadata(self, metadata_dict, tags_mapping):
-        tags_to_push = dict()
-        for key in tags_mapping.keys():
-            if key in metadata_dict:
-                value = metadata_dict[key]
-                if value.startswith('[') and value.endswith(']'):
-                    self.log_processing_info('Processing array value')
-                    value = value[1:-1]
-                    values = list(set(value.split(COMMA)))
-                    if len(values) != 1:
-                        self.log_processing_info('Empty or multiple metadata values, skipping [{}]'
-                                                 .format(key))
-                        continue
-                    value = values[0]
-                target_tag = tags_mapping[key]
-                if target_tag in tags_to_push:
-                    tags_to_push[target_tag].add(value)
-                else:
-                    tags_to_push[target_tag] = {value}
-        return tags_to_push
-
-    def map_tags(self, existing_attributes_dictionary):
-        tags_mapping_rules_str = os.getenv('HCS_PARSING_TAG_MAPPING', '')
-        tags_mapping_rules = tags_mapping_rules_str.split(COMMA) if tags_mapping_rules_str else []
-        skip_dictionary_check_tags = set(element.strip()
-                                         for element
-                                         in [os.getenv('HCS_PARSING_SKIP_DICTIONARY_CHECK_TAGS', '').split(COMMA)])
-        tags_mapping = dict()
-        for rule in tags_mapping_rules:
-            rule_mapping = rule.split(TAGS_MAPPING_KEYS_DELIMITER, 1)
-            if len(rule_mapping) != 2:
-                self.log_processing_info('Error [{}]: mapping rule declaration should contain a delimiter!'.format(
-                    rule_mapping))
-                continue
-            else:
-                key = rule_mapping[0]
-                value = rule_mapping[1]
-                if value not in existing_attributes_dictionary and value not in skip_dictionary_check_tags:
-                    self.log_processing_info('No dictionary [{}] is registered, the rule "{}" will be skipped!'
-                                             .format(value, rule_mapping))
-                    continue
-                else:
-                    tags_mapping[key] = value
-        return tags_mapping
-
-    def load_existing_attributes(self):
-        existing_attributes = self.api.execute_request(self.system_dictionaries_url)
-        existing_attributes_dictionary = {attribute['key']: attribute['values'] for attribute in existing_attributes}
-        return existing_attributes_dictionary
 
 
 class HcsFileParser:
@@ -511,9 +146,9 @@ class HcsFileParser:
     @staticmethod
     def calculate_wells_padding_for_ome(hcs_xml_info_root, ome_xml_info_root):
         wells_x_padding_hcs, \
-            wells_y_padding_hcs = HcsFileParser.extract_first_well_coordinates_hcs_xml(hcs_xml_info_root)
+        wells_y_padding_hcs = HcsFileParser.extract_first_well_coordinates_hcs_xml(hcs_xml_info_root)
         wells_x_padding_ome, \
-            wells_y_padding_ome = HcsFileParser.extract_first_well_coordinates_ome_xml(ome_xml_info_root)
+        wells_y_padding_ome = HcsFileParser.extract_first_well_coordinates_ome_xml(ome_xml_info_root)
         return wells_x_padding_hcs - wells_x_padding_ome, wells_y_padding_hcs - wells_y_padding_ome
 
     @staticmethod
@@ -1104,62 +739,3 @@ class HcsFileParser:
                     entry.find(hcs_schema_prefix + 'ImageResolutionY').text = \
                         str(float(resolution_y) * y_scaling).upper()
         return channel_dimensions
-
-
-def try_process_hcs(hcs_root_dir):
-    parser = None
-    processing_result = 1
-    try:
-        parser = HcsFileParser(hcs_root_dir)
-        processing_result = parser.process_file()
-        return processing_result
-    except Exception as e:
-        log_info('An error occurred during [{}] parsing: {}'.format(hcs_root_dir, str(e)))
-        print(traceback.format_exc())
-    finally:
-        if parser:
-            if processing_result != 2:
-                parser.clear_tmp_stat_file()
-            parser.clear_tmp_local_dir()
-
-
-def remove_not_empty_string(string_list):
-    return filter(lambda string: string is not None and len(string.strip()) > 0, string_list)
-
-
-def process_hcs_files():
-    paths_to_hcs_roots = remove_not_empty_string(os.getenv('HCS_TARGET_DIRECTORIES', '').split(COMMA))
-    if len(paths_to_hcs_roots) == 0:
-        lookup_paths = remove_not_empty_string(os.getenv('HCS_LOOKUP_DIRECTORIES', '').split(COMMA))
-        if not lookup_paths:
-            log_success('No paths for HCS processing specified')
-            exit(0)
-        log_info('Following paths are specified for processing: {}'.format(lookup_paths))
-        log_info('Lookup for unprocessed files')
-        paths_to_hcs_roots = HcsProcessingDirsGenerator(lookup_paths).generate_paths()
-        if not paths_to_hcs_roots:
-            log_success('Found no files requires processing in the lookup directories.')
-            exit(0)
-        log_info('Found {} files for processing.'.format(len(paths_to_hcs_roots)))
-    else:
-        log_info('{} files for processing are specified.'.format(len(paths_to_hcs_roots)))
-    processing_threads = int(os.getenv('HCS_PARSER_PROCESSING_THREADS', 1))
-    if processing_threads < 1:
-        log_info('Invalid number of threads [{}] is specified for processing, use single one instead'
-                 .format(processing_threads))
-        processing_threads = 1
-    log_info('{} thread(s) enabled for HCS processing'.format(processing_threads))
-    if TAGS_PROCESSING_ONLY:
-        log_info('Only tags will be processed, since TAGS_PROCESSING_ONLY is set to `true`')
-    if processing_threads == 1:
-        for file_path in paths_to_hcs_roots:
-            try_process_hcs(file_path)
-    else:
-        pool = multiprocessing.Pool(processing_threads)
-        pool.map(try_process_hcs, paths_to_hcs_roots)
-    log_success('Finished HCS files processing')
-    exit(0)
-
-
-if __name__ == '__main__':
-    process_hcs_files()
